feat: libft headers bonus segregation
